apps/python-api/src/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.27.1
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
azure-cosmos==4.7.0
passlib[bcrypt]==1.7.4
